+++ dialogy/base/plugin/__init__.py
@@ -183,25 +183,20 @@ Guard = Callable[[Input, Output], bool]
 
 class Plugin(ABC):
     """
-    Abstract class to be implemented by all plugins.
+        Abstract class to be implemented by all plugins.
 
-<<<<<<< HEAD
-=======
-    .. _plugin-abstract:
-
->>>>>>> origin/master
-    :param input_column: Transforms data in this column for a given dataframe, defaults to const.ALTERNATIVES
-    :type input_column: str
-    :param output_column: Saves transformation in this column for a given dataframe, defaults to None
-    :type output_column: Optional[str]
-    :param use_transform: Should the transformation be applied while training?, defaults to False
-    :type use_transform: bool
-    :param dest: The path where plugin output should be saved., defaults to None
-    :type dest: Optional[str]
-    :param guards: A list of functions that evaluate to bool, defaults to None
-    :type guards: Optional[List[Guard]]
-    :param debug: Should we print debug logs?, defaults to False
-    :type debug: bool, optional
+        :param input_column: Transforms data in this column for a given dataframe, defaults to const.ALTERNATIVES
+        :type input_column: str
+        :param output_column: Saves transformation in this column for a given dataframe, defaults to None
+        :type output_column: Optional[str]
+        :param use_transform: Should the transformation be applied while training?, defaults to False
+        :type use_transform: bool
+        :param dest: The path where plugin output should be saved., defaults to None
+        :type dest: Optional[str]
+        :param guards: A list of functions that evaluate to bool, defaults to None
+        :type guards: Optional[List[Guard]]
+        :param debug: Should we print debug logs?, defaults to False
+        :type debug: bool, optional
     """
 
     def __init__(
@@ -235,7 +230,6 @@ class Plugin(ABC):
 
     def __call__(self, workflow: Workflow) -> None:
         """
-<<<<<<< HEAD
         Set workflow with output.
 
         This important book-keeping method is called by the workflow.
@@ -247,9 +241,6 @@ class Plugin(ABC):
         - The workflow takes care of keeping its :ref:`input<Input>` and :ref:`output<Output>` immutable.
 
         .. _PluginBookkeeping:
-=======
-        Use the plugin's logic and save the result within the workflow.
->>>>>>> origin/master
 
         :param workflow: An instance of :ref:`Workflow <WorkflowClass>`.
         :type workflow: Workflow
@@ -273,13 +264,10 @@ class Plugin(ABC):
         """
         Decide if the plugin should execute.
 
-<<<<<<< HEAD
         If this method returns true, the plugin's utility method will not be called.
 
         .. _Guards:
 
-=======
->>>>>>> origin/master
         :return: prevent plugin execution if True.
         :rtype: bool
         """