+++ dialogy/constants/__init__.py
@@ -36,6 +36,9 @@ AMOUNT_OF_MONEY = "amount-of-money"
 DURATION = "duration"
 CREDIT_CARD_NUMBER = "credit-card-number"
 
+FUTURE = "future"
+PAST = "past"
+ANY = "__any__"
 
 DUCKLING_DIMS = [PEOPLE, NUMBER, AMOUNT_OF_MONEY, DURATION, CREDIT_CARD_NUMBER, TIME]