+++ dialogy/utils/datetime.py
@@ -72,3 +72,7 @@ def make_unix_ts(tz: str = "UTC") -> Callable[[str], int]:
         return dt2timestamp(dt)
 
     return make_tz_aware
+
+
+def unix_ts_to_datetime(reference_time: int, timezone: str = "UTC") -> datetime:
+    return datetime.fromtimestamp(reference_time / 1000, pytz.timezone(timezone))